+++ src/com/pinktwins/elephant/Elephant.java
@@ -10,17 +10,19 @@ import java.util.logging.Logger;
 import javax.swing.UIManager;
 
 import com.google.common.eventbus.EventBus;
+import com.pinktwins.elephant.data.ElephantUndoManager;
 import com.pinktwins.elephant.data.Settings;
 import com.pinktwins.elephant.data.Vault;
 
 public class Elephant {
 
-	public static final int VERSION = 31;
+	public static final int VERSION = 43;
 
 	private static final Logger LOG = Logger.getLogger(Elephant.class.getName());
 
 	public static final EventBus eventBus = new EventBus();
 	public static final Settings settings = new Settings();
+	public static final ElephantUndoManager undoManager = new ElephantUndoManager();
 
 	public static String[] args;