+++ src/com/pinktwins/elephant/Start.java
@@ -16,6 +16,7 @@ import javax.swing.JButton;
 import javax.swing.JFileChooser;
 import javax.swing.JLabel;
 import javax.swing.JPanel;
+import javax.swing.JCheckBox;
 
 import com.pinktwins.elephant.data.Note;
 import com.pinktwins.elephant.data.Vault;
@@ -40,8 +41,8 @@ public class Start extends BackgroundPanel {
 
 		setLayout(new FlowLayout());
 
-		JPanel main = new JPanel(new GridLayout(3, 1));
-		main.setBorder(BorderFactory.createEmptyBorder(200, 0, 0, 0));
+		JPanel main = new JPanel(new GridLayout(4, 1));
+		main.setBorder(BorderFactory.createEmptyBorder(140, 0, 0, 0));
 
 		JLabel welcome = new JLabel("Please choose your note location.", JLabel.CENTER);
 		welcome.setForeground(Color.DARK_GRAY);
@@ -50,16 +51,34 @@ public class Start extends BackgroundPanel {
 
 		JButton bLocation = new JButton("Choose folder");
 
-		JLabel hint = new JLabel("Folder 'Elephant' will be created under this folder.", JLabel.CENTER);
+		final JCheckBox createElephantFolder = new JCheckBox("Create folder 'Elephant' under this location.");
+		createElephantFolder.setForeground(Color.DARK_GRAY);
+		createElephantFolder.setFont(ElephantWindow.fontStart);
+		createElephantFolder.setBorder(BorderFactory.createEmptyBorder(20, 0, 0, 0));
+		createElephantFolder.setSelected(true);
+		
+		final JLabel hint = new JLabel("", JLabel.CENTER);
 		hint.setForeground(Color.DARK_GRAY);
 		hint.setFont(ElephantWindow.fontStart);
 		hint.setBorder(BorderFactory.createEmptyBorder(20, 0, 0, 0));
 
 		main.add(welcome);
 		main.add(bLocation);
+		main.add(createElephantFolder);
 		main.add(hint);
 
 		add(main);
+
+		createElephantFolder.addActionListener(new ActionListener() {
+			@Override
+			public void actionPerformed(ActionEvent e) {
+				if (createElephantFolder.isSelected()) {
+					hint.setText("");
+				} else {
+					hint.setText("Elephant will use the selected folder directly.");
+				}
+			}
+		});
 
 		bLocation.addActionListener(new ActionListener() {
 			@Override
@@ -73,7 +92,13 @@ public class Start extends BackgroundPanel {
 				if (res == JFileChooser.APPROVE_OPTION) {
 					File f = ch.getSelectedFile();
 					if (f.exists()) {
-						File folder = new File(f + File.separator + "Elephant");
+						File folder = null;
+						if (createElephantFolder.isSelected()) {
+							folder = new File(f.getAbsolutePath() + File.separator + "Elephant");
+						} else {
+							folder = new File(f.getAbsolutePath());
+						}
+
 						if (folder.exists() || folder.mkdirs()) {
 
 							Vault.getInstance().setLocation(folder.getAbsolutePath());