+++ src/com/pinktwins/elephant/data/Settings.java
@@ -2,11 +2,13 @@ package com.pinktwins.elephant.data;
 
 import java.io.File;
 import java.io.IOException;
+import java.nio.charset.Charset;
 import java.util.logging.Logger;
 
 import org.json.JSONException;
 import org.json.JSONObject;
 
+import com.google.common.io.Files;
 import com.pinktwins.elephant.NoteList.ListModes;
 import com.pinktwins.elephant.Sidebar.RecentNotesModes;
 import com.pinktwins.elephant.util.IOUtil;
@@ -16,9 +18,10 @@ public class Settings {
 	private static final Logger LOG = Logger.getLogger(Settings.class.getName());
 
 	public static enum Keys {
-		DEFAULT_NOTEBOOK("defaultNotebook"), VAULT_FOLDER("noteFolder"), USE_LUCENE("useLucene"), NOTELIST_MODE("noteListMode"), AUTOBULLET("autoBullet"), RECENT_SHOW(
-				"showRecent"), ALLOW_FILENAMECHARS("allowFilenameChars"), CONFIRM_DELETE_FROM_TRASH("confirmDeleteFromTrash"), WINDOW_MAXIMIZED("maximized"), FONT_SCALE(
-				"fontScale"), PASTE_PLAINTEXT("pastePlaintext"), SHOW_SIDEBAR("showSidebar");
+		DEFAULT_NOTEBOOK("defaultNotebook"), VAULT_FOLDER("noteFolder"), USE_LUCENE("useLucene"), NOTELIST_MODE("noteListMode"), AUTOBULLET(
+				"autoBullet"), RECENT_SHOW("showRecent"), ALLOW_FILENAMECHARS("allowFilenameChars"), CONFIRM_DELETE_FROM_TRASH(
+						"confirmDeleteFromTrash"), WINDOW_MAXIMIZED("maximized"), FONT_SCALE("fontScale"), PASTE_PLAINTEXT(
+								"pastePlaintext"), SHOW_SIDEBAR("showSidebar"), DEFAULT_FILETYPE("defaultFiletype"), CHARSET("charset");
 
 		private final String str;
 
@@ -152,7 +155,7 @@ public class Settings {
 
 	private void save() {
 		try {
-			IOUtil.writeFile(settingsFile(), map.toString(4));
+			Files.write(map.toString(4), settingsFile(), Charset.forName("UTF-8"));
 		} catch (IOException e) {
 			LOG.severe("Fail: " + e);
 		} catch (JSONException e) {
@@ -224,5 +227,24 @@ public class Settings {
 			return true;
 		}
 		return getBoolean(Keys.SHOW_SIDEBAR);
+	}
+
+	public String getDefaultFiletype() {
+		if (!has(Keys.DEFAULT_FILETYPE)) {
+			return "txt";
+		}
+		return getString(Keys.DEFAULT_FILETYPE);
+	}
+
+	public boolean hasCharset() {
+		return has(Keys.CHARSET);
+	}
+
+	public String getCharset() {
+		if (hasCharset()) {
+			return getString(Keys.CHARSET);
+		} else {
+			return null;
+		}
 	}
 }